requirements.txt
hydra-core==1.3.2
pandas==2.1.4
patsy==0.5.6
scikit-learn==1.3.2
hyperopt==0.2.7
xgboost==2.0.3
mlflow==2.9.2
dagshub==0.3.12
deepchecks==0.17.5
pandera==0.18.0
pytest-steps==1.8.0
dvc==3.38.1
bentoml==1.1.11
streamlit==1.29.0
black==23.12.1